+++ src/components/chips/filter_chip/FilterChip.stories.tsx
@@ -138,7 +138,7 @@ function getProps(item: Item): FilterChipProps & PseudoProps {
     case "enabled":
       return base;
     case "disabled":
-      return { ...base, disabled: true };
+      return { ...base, state: "disabled" };
     case "focused":
       return { ...base, "data-focus-visible": "" };
     case "hovered":