+++ src/components/chips/input_chip/InputChip.stories.tsx
@@ -120,7 +120,7 @@ function getProps(item: Item): InputChipProps & PseudoProps {
     case "enabled":
       return base;
     case "disabled":
-      return { ...base, disabled: true };
+      return { ...base, state: "disabled" };
     case "focused":
       return { ...base, "data-focus-visible": "" };
     case "hovered":