+++ src/mod.ts
@@ -13,3 +13,4 @@ export * from "./components/divider/mod.ts";
 export * from "./components/lists/mod.ts";
 export * from "./components/fab/mod.ts";
 export * from "./components/tokens/mod.ts";
+export * from "./components/types.ts";